requirements.txt
loguru
requests
mysql-connector-python